Refactor sink task integration tests to be more readable (#543)
* wrong iterations

* fix it

* comments and todo

* double trouble is truely trouble

* multi logs work, still need client mapping test

* all pass

* format and quick comment

* tests pass, need reset method

* no sunny days in seattle

* move to diff file

* formatttttt

* ignore logging test bc jdbc flakiness
